Set up node_modules for compiled code, and put compiled code in .pyret/
With the addition of libraries like `node-canvas`, which rely on
platform-specific FFIs and C code compilation, we cannot have CLI Pyret only
generate standalone JS files: they necessarily require some node imports with
`require` to function.

Thankfully, we can list these required packages in the `pyret-npm` repo, and
add a link to those `node_modules` in the `.pyret` directory where other
compiled files go.

In order to make this resolve correctly, the generated code that's run needs to
be adjacent to or in a subdirectory relative to `node_modules`. So this now
sets up that:

$ pyret my-file.arr

creates this structure:

.pyret/
  my-file.jarr # standalone, run automatically by pyret if compilation succeeds
               # also runnable with node .pyret/my-file.jarr
  compiled/ # standard cached compiled files
  node_modules@ # a symlink to <user npm dir>/pyret-npm/node_modules

This way, anything pyret needs is installed along with `pyret-npm`, and can be
referred to  appropriately from `my-file.jarr`.

In addition, because of how `node_modules` lookup semantics works, if the
programmer has their _own_ `node_modules` that their Pyret program relies on,
node will still find it in the parent directory of `.pyret`